requirements.txt
spidev
RPi.GPIO
Pillow
numpy
